refactor(tracker): extract path resolution, lock, atomic write into tracker-utils.mjs
Move writeFileAtomic, canonicalizeTrackerPath, tracker-path resolution, and
the tracker lock (acquireTrackerLock + stale-recovery helpers) out of
merge-tracker.mjs into the shared tracker-utils.mjs, and add states.yml
loading (loadCanonicalStates) with a strict alias-aware resolver
(resolveCanonicalState).

Groundwork for set-status.mjs (#1428): every tracker writer must share the
same canonical path and the same lock so concurrent writers exclude each
other instead of last-writer-wins. Behavior of merge-tracker is unchanged
(same lock dir naming, same env overrides).
